only manage it we actually have a webuser

+++ manifests/repostorage.pp
@@ -153,11 +153,13 @@ define gitosis::repostorage(
     context => "/files/etc/group",
   }
 
-  git::web::repo{$git_vhost:
-    logmode => $logmode,
+  $webuser = hiera('gitweb_webserver','none')
+  if $webuser != 'none' {
+    git::web::repo{$git_vhost:
+      logmode => $logmode,
+    }
   }
 
-  $webuser = hiera('gitweb_webserver','none')
   if $gitweb and $ensure == 'present' {
     case $git_vhost {
       'absent': { fail("can't do gitweb if \$git_vhost isn't set for ${name} on ${::fqdn}") }
@@ -198,7 +200,7 @@ define gitosis::repostorage(
       ensure => 'absent',
     }
     # if present is absent we removed the user anyway
-    if ($present != 'absent') {
+    if ($present != 'absent') and ($webuser != 'absent'){
       Augeas["manage_webuser_in_repos_group_${real_group_name}"]{
         changes => "rm ${real_group_name}/user[.='${webuser}']",
       }
